+++ src/api/v1/otp/__test__/otp.validation.test.ts
@@ -1,49 +1,34 @@
 import { ZodError } from "zod";
 
-import { EmailEventType } from "../../../../services";
+import { NotificationType } from "@/constants/events.constant";
+
 import { OtpEvent, otpEvent } from "../otp.validation";
 
 describe("Test OTP validation", () => {
   const otpTypes: OtpEvent = {
-    otpType: EmailEventType.sendEmailVerificationOTP,
+    otpType: NotificationType.EMAIL_VERIFICATION_OTP,
     email: "example@gmail.com",
   };
 
-  it("should validate sendOtpSchema", () => {
-    const res = otpEvent.safeParse(otpTypes);
-    expect(res.success).toBe(true);
-  });
-
-  it(`should throw error if email is not provided with otpType of ${EmailEventType.sendEmailVerificationOTP}`, () => {
+  it(`should throw error if email is not provided with otpType of ${NotificationType.EMAIL_VERIFICATION_OTP}`, () => {
     const otpTypesWithoutEmail = { ...otpTypes, email: undefined };
 
     try {
       otpEvent.parse(otpTypesWithoutEmail);
     } catch (error) {
-      expect((error as ZodError).errors[0].message).toBe("Email is required");
+      expect((error as ZodError).errors[0].message).toBe("Required");
+      expect((error as ZodError).errors[0].path).toEqual(["email"]);
     }
   });
 
-  it("should throw error if phone number is not provided", () => {
-    const otpTypesWithoutPhoneNumber = { otpType: "verifyPhoneNumber" };
+  // it("should throw error if phone number is not provided", () => {
+  //   const otpTypesWithoutPhoneNumber = { otpType: NotificationType.PHONE_VERIFICATION };
 
-    try {
-      otpEvent.parse(otpTypesWithoutPhoneNumber);
-    } catch (error) {
-      expect((error as ZodError).errors[0].message).toBe("Phone number is required");
-    }
-  });
-
-  it("should throw error if phone number is invalid", () => {
-    const otpTypesWithInvalidPhoneNumber = {
-      otpType: "verifyPhoneNumber",
-      phoneNumber: "1234567890",
-    };
-
-    try {
-      otpEvent.parse(otpTypesWithInvalidPhoneNumber);
-    } catch (error) {
-      expect((error as ZodError).errors[0].message).toBe("Invalid Phone number");
-    }
-  });
+  //   try {
+  //     otpEvent.parse(otpTypesWithoutPhoneNumber);
+  //   } catch (error) {
+  //     expect((error as ZodError).errors[0].message).toBe("Required");
+  //     expect((error as ZodError).errors[0].path).toEqual(["phoneNumber"]);
+  //   }
+  // });
 });