+++ src/utils/index.ts
@@ -1,6 +1,6 @@
 export * from "./helper.util";
 export * from "./jwt.util";
-export { default as logger } from "./logger";
+export * from "./logger";
 export * from "./otp.util";
 export * from "./password.util";
 export * from "./send-response.util";