+++ src/utils/jwt.util.ts
@@ -8,10 +8,17 @@ import {
   jwtActionTokenSchema,
   jwtRefreshTokenSchema,
 } from "@/api/v1/auth/auth.validation";
-import { CURRENT_SERVICE, ErrorTypeEnum, Permission, ServiceEnum, envConstants } from "@/constants";
+import {
+  ACTION_TOKEN_EXPIRY_TIME,
+  CURRENT_SERVICE,
+  ErrorTypeEnum,
+  Permission,
+  ServiceEnum,
+  envConstants,
+} from "@/constants";
 
-import { logger } from ".";
 import { CryptoUtil } from "./crypto.util";
+import { errorLogger } from "./logger";
 
 const { JWT_ACCESS_SECRET, JWT_REFRESH_SECRET, JWT_TOKEN_FOR_ACTION_SECRET } = envConstants;
 
@@ -42,7 +49,7 @@ export const generateAccessToken = async (payload: { userId: string; permissions
       issuer: CURRENT_SERVICE,
     });
   } catch (error) {
-    logger.error(`Access token generation error: ${error}`);
+    errorLogger.error(`Access token generation error: ${error}`);
     throw new Error(ErrorTypeEnum.enum.INTERNAL_SERVER_ERROR);
   }
 };
@@ -63,7 +70,7 @@ export const generateRefreshToken = async (payload: JwtRefreshToken): Promise<st
 export const generateTokenForAction = (payload: JwtActionToken) => {
   const validPayload = jwtActionTokenSchema.parse(payload);
   return jwt.sign(validPayload, JWT_TOKEN_FOR_ACTION_SECRET, {
-    expiresIn: "5m",
+    expiresIn: ACTION_TOKEN_EXPIRY_TIME,
     audience: CURRENT_SERVICE,
     issuer: CURRENT_SERVICE,
   });