+++ src/utils/send-response.util.ts
@@ -1,38 +1,67 @@
-import { Response, request } from "express";
+import { Response } from "express";
 import { ZodIssue } from "zod";
 
-import { envConstants } from "@/constants";
+import { ErrorCode } from "@/constants";
 
-import { errorLogger } from "./logger";
-
-export interface SendResponse<T = undefined> {
+export interface SendResponseSuccess<T> {
   response: Response;
-  code?: string;
   statusCode: number;
-  status?: "success" | "failed";
+  status?: "success";
   message: string;
-  errorDetails?: Error;
   data?: T;
+}
+
+export interface SendResponseFailed {
+  response: Response;
+  statusCode: number;
+  status: "failed";
+  message: string;
+  code: ErrorCode;
+  errors: ZodIssue[];
+}
+
+export type SendResponse<T> = SendResponseSuccess<T> | SendResponseFailed;
+
+interface IResponseBodySuccess<T> {
+  status: "success";
+  message: string;
+  data?: T;
+}
+
+interface IResponseBodyFailed {
+  status: "failed";
+  message: string;
+  code: string;
   errors?: ZodIssue[];
 }
 
-export const sendResponse = <T>(responseData: SendResponse<T>) => {
-  const { response, statusCode, message, errorDetails, status = "success", data, errors, code } = responseData;
-  const isProduction = envConstants.NODE_ENV === "production";
+type IResponseBody<T> = IResponseBodySuccess<T> | IResponseBodyFailed;
 
-  const responseBody: Record<string, unknown> = { status, message, data, code, errors };
+export const sendResponse = <T>({ status = "success", message, response, statusCode, ...rest }: SendResponse<T>) => {
+  let responseBody: IResponseBody<T>;
 
-  if (!isProduction && errorDetails) {
-    responseBody.errorDetails = {
-      message: errorDetails.message,
-      stack: errorDetails.stack,
+  if (status === "failed") {
+    const { code, errors } = rest as SendResponseFailed;
+
+    // Initialize response body based on status
+    responseBody = {
+      status,
+      message,
+      code,
     };
-  }
 
-  if (errorDetails && statusCode >= 500) {
-    errorLogger.error(
-      `Error occurred in ${request.url}: ${errorDetails.name} ${errorDetails.message} ${errorDetails.stack}`
-    );
+    if (errors.length) {
+      responseBody.errors = errors;
+    }
+  } else {
+    const { data } = rest as SendResponseSuccess<T>;
+
+    // Initialize response body based on status
+    responseBody = {
+      status,
+      message,
+      data,
+    };
   }
 
   response.status(statusCode).json(responseBody);