+++ src/utils/test/auth.utils.ts
@@ -1,14 +1,13 @@
 import supertest, { Response } from "supertest";
 
 import { success } from "@/api/v1/auth/auth.constant";
-import { Login } from "@/api/v1/auth/auth.validation";
+import { Login, SignUpResponse } from "@/api/v1/auth/auth.validation";
 import { app } from "@/app";
 import { ErrorTypeEnum, STATUS_CODES, errorMap } from "@/constants";
+import { NotificationType } from "@/constants/events.constant";
 
-import { CreateUser, ResetPassword } from "../../api/v1/user/user.validation";
-import { EmailEventType } from "../../services";
-import { expectOTPRequestSuccess, expectOTPVerificationSuccess, requestOTP, retrieveOTP, verifyOTP } from "./otp.utils";
-import { expectFindUserByUsernameSuccess, findUserByUsername } from "./user.utils";
+import { ResetPassword } from "../../api/v1/user/user.validation";
+import { expectOTPVerificationSuccess, retrieveOTP, verifyOTP } from "./otp.utils";
 
 export const login = async (loginData: Login): Promise<Response> => {
   return supertest(app).post("/api/v1/auth/login").send(loginData);
@@ -22,13 +21,15 @@ export const expectLoginSuccess = (response: Response): void => {
   const authorizationHeader = headers["authorization"];
   expect(authorizationHeader).toBeDefined();
 
-  const setCookieHeader = response.get("set-cookie")?.[0];
-  expect(setCookieHeader).toContain("refresh-token=");
-  expect(setCookieHeader).toMatch(/HttpOnly; Secure/);
+  const refreshToken = headers["refresh-token"];
+  expect(refreshToken).toBeDefined();
 
   expect(body).toMatchObject({
     message: success.LOGGED_IN_SUCCESSFULLY,
     status: "success",
+    data: {
+      userId: expect.any(String),
+    },
   });
 };
 
@@ -56,6 +57,10 @@ export const expectSignUpSuccess = (response: Response): void => {
 
   expect(body).toMatchObject({
     message: success.SIGN_UP_SUCCESS,
+    data: {
+      token: expect.any(String),
+      userId: expect.any(String),
+    },
   });
 };
 
@@ -80,31 +85,33 @@ export const expectRenewTokenSuccess = (response: Response) => {
 
   expect(statusCode).toBe(STATUS_CODES.OK);
 
-  const newRefreshToken = headers["authorization"];
-  expect(newRefreshToken).toBeDefined();
+  const authorizationHeader = headers["authorization"];
+  expect(authorizationHeader).toBeDefined();
+
+  const refreshToken = headers["refresh-token"];
+  expect(refreshToken).toBeDefined();
 
   expect(response.body).toMatchObject({
     message: success.TOKEN_RENEWED_SUCCESSFULLY,
     status: "success",
+    data: {
+      userId: expect.any(String),
+    },
   });
 };
 
-export const verifyAccount = async (user: CreateUser) => {
-  // Step 1: Request OTP
-  const { email, username } = user;
-  const otpResponse = await requestOTP(email);
-  expectOTPRequestSuccess(otpResponse);
+export const verifyAccount = async ({ userId, token }: SignUpResponse) => {
+  const otpType = NotificationType.EMAIL_VERIFICATION_OTP;
+  // Step 1: Retrieve OTP from database
+  const otpData = await retrieveOTP(userId, otpType);
 
-  // Step 2: Retrieve User from database
-  const userResponse = await findUserByUsername(username);
-  expectFindUserByUsernameSuccess(userResponse, user);
-
-  // Step 2: Retrieve OTP from database
-  const otpData = await retrieveOTP(userResponse.body.data.id, EmailEventType.sendEmailVerificationOTP);
-
-  // Step 3: Verify OTP
-  const verifyResponse = await verifyOTP(otpData, email);
-  expectOTPVerificationSuccess(verifyResponse);
+  // Step 2: Verify OTP
+  const verifyResponse = await verifyOTP({
+    otp: otpData.otp,
+    token: token,
+    otpType,
+  });
+  expectOTPVerificationSuccess(otpType, verifyResponse);
 };
 
 export const forgetPassword = async (email: string): Promise<Response> => {
@@ -116,7 +123,10 @@ export const expectForgetPasswordSuccess = (response: Response): void => {
   const { statusCode, body } = response;
 
   expect(statusCode).toBe(STATUS_CODES.OK);
-  expect(body).toMatchObject({ message: success.FORGET_PASSWORD_EMAIL_SENT });
+  expect(body).toMatchObject({
+    message: success.FORGET_PASSWORD_EMAIL_SENT,
+    data: { token: expect.any(String) },
+  });
 };
 
 export const resetPassword = async (resetPassword: ResetPassword): Promise<Response> => {
@@ -124,9 +134,21 @@ export const resetPassword = async (resetPassword: ResetPassword): Promise<Respo
 };
 
 export const expectResetPasswordSuccess = (response: Response): void => {
-  expect(response).toBeDefined();
-  const { statusCode, body } = response;
+  const { statusCode, headers } = response;
 
   expect(statusCode).toBe(STATUS_CODES.OK);
-  expect(body).toMatchObject({ message: success.PASSWORD_RESET_SUCCESSFULLY });
+
+  const authorizationHeader = headers["authorization"];
+  expect(authorizationHeader).toBeDefined();
+
+  const refreshToken = headers["refresh-token"];
+  expect(refreshToken).toBeDefined();
+
+  expect(response.body).toMatchObject({
+    message: success.PASSWORD_RESET_SUCCESSFULLY,
+    status: "success",
+    data: {
+      userId: expect.any(String),
+    },
+  });
 };